Move link into its own file
